Update Modal with shadow and use XAML
