+++ Trimble.Modus.Components/Constant/ColorsConstants.cs
@@ -28,4 +28,5 @@ internal class ColorsConstants
     public const string SuccessToastColor = "SuccessToastColor";
     public const string ToastTextBlue = "ToastTextBlue";
     public const string ModalGraySemiTransparent = "ModalGraySemiTransparent";
+    public const string ModalShadowColor = "ModalShadowColor";
 }
