+++ Trimble.Modus.Components/Trimble.Modus.Components.csproj
@@ -75,6 +75,9 @@
 	  <Compile Update="Controls\TMInput\TMInput.xaml.cs">
 	    <DependentUpon>TMInput.xaml</DependentUpon>
 	  </Compile>
+	  <Compile Update="Controls\TMModal\TMModalContents.xaml.cs">
+	    <DependentUpon>TMModalContents.xaml</DependentUpon>
+	  </Compile>
 	  <Compile Update="Handlers\TMButtonAndroidTouchHandler.cs">
 	    <ExcludeFromCurrentConfiguration>false</ExcludeFromCurrentConfiguration>
 	  </Compile>
@@ -89,6 +92,9 @@
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	  <MauiXaml Update="Controls\TMInput\TMInput.xaml">
+	    <Generator>MSBuild:Compile</Generator>
+	  </MauiXaml>
+	  <MauiXaml Update="Controls\TMModal\TMModalContents.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
 	  <MauiXaml Update="Controls\TMNumberInput\TMNumberInput.xaml">
